+++ src/commands/version/version-set.ts
@@ -1,8 +1,7 @@
-import { TerminalInterface } from "cli/terminalInterface";
+import { TerminalInterface } from "cli/manager/terminalInterface";
 import { CommandRegistrar } from "commands/command-registrar";
 import { Prompt } from "utils/prompt-utils";
 import * as vscode from 'vscode';
-import { EnvironmentRefresher } from '../../utils/vscode-utils';
 export class versionSetCommand extends CommandRegistrar {
     protected async execute(...args: any[]): Promise<void> {
         const tag: String | undefined = await Prompt.Text("Pls add your version tag (leave empty for selection)","v5.0.0-beta6");
